requirements.txt
feedparser==6.0.10
requests==2.31.0
google-generativeai==0.8.0
pyyaml==6.0.1
beautifulsoup4==4.12.2
python-dateutil==2.8.2